requirements.txt
torch>=1.13.1
transformers==4.21.2
ujson
requests
flask
pyjnius
pyserini==0.18.0
faiss-cpu==1.7.2
